+++ src/config/Progress.ts
@@ -1,0 +1,12 @@
+interface Progress {
+  total_num: Number;
+  now_num: Number;
+  avisynth_flag: boolean;
+  percent: Number;
+  log_updated: boolean;
+  log: string;
+  duration: Number;
+  steps: Number;
+  step: Number;
+}
+export { Progress };
